finished final project
Fixed most of the mechanic that I did not do in the journals, specifically the gravity as I could not get it to work before, however now it does. Implemented all of the mechanics necessary that I wanted to. Dash is now a ground pound and you can wall jump and hold on